Hoje eu arrumei todos os CRUDS do projeto e deixei no padrão que o Diego F. me auxiliou. Todos os controllers e services eu criei na mão, sem o comando que cria automaticamente, e creio que me ajudou bastante a entender mais sobre os padrões do rails.

+++ modulo_05/time_to_answer/app/services/admins_backoffice/admins_service.rb
@@ -10,11 +10,9 @@ class AdminsBackoffice::AdminsService
   end
 
   def self.update(params)
-    puts "Params dentro do service #{params}"
-    admin = Admin.where(id: params[:id]).first || Admin.new
+    admin = Admin.where(id: params[:id]).first
 
     admin.assign_attributes(params)
-
 
     if admin.save
       [:success, notice: "Administrador atualizado com sucesso"]
@@ -24,15 +22,15 @@ class AdminsBackoffice::AdminsService
   end
 
   def self.destroy(params)
-
     admin = Admin.where(id: params[:id]).first
-
-    admin.assign_attributes(params)
-
-    if admin.destroy
-      [:success, notice: "Administrador excluído com sucesso"]
-    else
+    if admin.blank?
       [:error, admin]
+    else
+      if admin.destroy
+        [:success, notice: "Administrador excluído com sucesso"]
+      else
+        [:error, admin]
+      end
     end
   end
